+++ src/modules/activities/index.ts
@@ -1,0 +1,3 @@
+export { default as ActivitiesPage } from './ActivitiesPage';
+export { default as NewActivityPage } from './NewActivityPage';
+export * from './activities.types';